+++ UnitTests/UnitTest1.cs
@@ -13,22 +13,22 @@ namespace UnitTests
         public void Test1()
         {
             int numberOfClustersToTest = 5;
-            var embeddings = new double[][]
+            var embeddings = new float[][]
             {
-                new double[] { -0.0628880858, -0.0427831635},
-                new double[] { -0.0610538125, 0.0613973737},
-                new double[] { -0.06559628, -0.09377544},
-                new double[] { -0.0559743047, 0.06577249},
-                new double[] { -0.08143592, 0.000687024556 },
-                new double[] { -0.06379545, -0.0174551066 },
-                new double[] { -0.05620724, -0.012291 },
-                new double[] { -0.101768017, -0.3260213 },
-                new double[] { -0.07536936, 0.154612169 },
-                new double[] { -0.0763737, 0.1386943 },
-                new double[] { -0.0613845, 0.109149076 },
-                new double[] { -0.05336821, 0.07296084 },
-                new double[] { -0.0598036051, 0.0348322764 },
-                new double[] { -0.08613151, -0.153230369 }
+                new float[] { -0.0628880858f, -0.0427831635f},
+                new float[] { -0.0610538125f, 0.0613973737f},
+                new float[] { -0.06559628f, -0.09377544f},
+                new float[] { -0.0559743047f, 0.06577249f},
+                new float[] { -0.08143592f, 0.000687024556f },
+                new float[] { -0.06379545f, -0.0174551066f },
+                new float[] { -0.05620724f, -0.012291f },
+                new float[] { -0.101768017f, -0.3260213f },
+                new float[] { -0.07536936f, 0.154612169f },
+                new float[] { -0.0763737f, 0.1386943f },
+                new float[] { -0.0613845f, 0.109149076f },
+                new float[] { -0.05336821f, 0.07296084f },
+                new float[] { -0.0598036051f, 0.0348322764f },
+                new float[] { -0.08613151f, -0.153230369f }
             };
 
             var recommendedClusters = new ClusteringEngine().GetRecommendation(embeddings, numberOfClustersToTest);
